[🚚move] 일정관리 폴더 정리
