+++ src/stores/selectedDateSlice.tsx
@@ -1,7 +1,7 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 
 import { Schedules_list_counseling, Schedules_list_private } from '@apis/schedulesAPIs';
-import { CalendarEventType } from '@components/schedules/Calendar';
+import { CalendarEventType } from '@components/schedules/calendar/Calendar';
 import { getCurrentDate } from '@utils/getDate';
 
 const { month: curMonth, year: curYear, date: curDate } = getCurrentDate();
